versión mvc sw ok

+++ ecommerce_mvc/src/main/java/com/inditex/ecommerce/model/dto/ProductResponse.java
@@ -17,7 +17,7 @@ public class ProductResponse {
 
     private Long productId;
     private Long brandId;
-    private Integer priceList;
+    private Long priceList;
     private Double price;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
